minor bug fixes

+++ multicamera_airflow_pipeline/tim_240731/keypoints/predict_2D.py
@@ -1,13 +1,6 @@
-from joblib import Parallel, delayed
 import numpy as np
-from copy import deepcopy
 import sys
-import pandas as pd
 from pathlib import Path
-import networkx as nx
-import matplotlib.pyplot as plt
-import glob
-import joblib
 import os
 import h5py
 from tqdm.auto import tqdm
@@ -15,8 +8,6 @@ import logging
 import shutil
 import tempfile
 import cv2
-import re
-import subprocess
 from motpy import Detection, MultiObjectTracker
 from mmdet.apis import init_detector, inference_detector
 from mmpose.utils import adapt_mmdet_pipeline
@@ -24,16 +15,8 @@ from mmpose.apis import inference_topdown, init_model as init_pose_estimator
 from datetime import datetime
 import torch
 from datetime import datetime, timedelta
+import signal
 
-# from mmdeploy_runtime import Detector, PoseDetector
-from multicamera_airflow_pipeline.tim_240731.skeletons.defaults import (
-    dataset_info,
-    parents_dict,
-    keypoint_info,
-    keypoints,
-    keypoints_order,
-    kpt_dict,
-)
 logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
 logger.info(f"Python interpreter binary location: {sys.executable}")
@@ -164,7 +147,7 @@ class Inferencer2D:
 
             output_h5_file = self.output_directory_predictions / f"{video_path.stem}.h5"
             if output_h5_file.exists() and not self.recompute_completed:
-                logger.info(f"Completed, skipping {video_path}")
+                # logger.info(f"Completed, skipping {video_path}")
                 continue
 
             # if a log file exists, then this video is being processed by another job
@@ -175,44 +158,57 @@ class Inferencer2D:
                 with open(log_file, "r") as f:
                     log_time = datetime.strptime(f.readline().strip(), "%Y-%m-%d %H:%M:%S.%f")
                 if datetime.now() - log_time < timedelta(hours=6):
-                    logger.info(f"Log file exists, skipping {log_file}")
+                    logger.info(f"Log file exists from {log_time}, skipping {log_file}")
                     continue
 
             # initially save output to temp file
             with tempfile.NamedTemporaryFile(suffix=".h5", delete=False) as temp_h5_file:
                 temp_h5_path = temp_h5_file.name
-
                 # create a log with the name of this video, to indicate that it is being processed
                 #   the log will have a timestamp
                 with open(log_file, "w") as f:
                     f.write(f"{datetime.now()}\n")
 
-                predict_video(
-                    video_path=video_path,
-                    output_h5_file=temp_h5_path,
-                    detector=self.detector,
-                    pose_estimator=self.pose_estimator,
-                    n_keypoints=self.n_keypoints,
-                    detection_interval=self.detection_interval,
-                    n_animals=self.n_animals,
-                    use_motpy=self.use_motpy,
-                    n_motpy_tracks=self.n_motpy_tracks,
-                    use_tensorrt=self.use_tensorrt,
-                    total_frames=self.expected_video_length_frames,
-                )
-                # when completed, move to output file
-                shutil.copy(temp_h5_path, output_h5_file)
-                os.remove(temp_h5_path)
-        
-        
+                # update the signal handler for the current video log file
+                update_log_file_signal_handler(log_file)
+
+                try:
+                    # do prediction
+                    predict_video(
+                        video_path=video_path,
+                        output_h5_file=temp_h5_path,
+                        detector=self.detector,
+                        pose_estimator=self.pose_estimator,
+                        n_keypoints=self.n_keypoints,
+                        detection_interval=self.detection_interval,
+                        n_animals=self.n_animals,
+                        use_motpy=self.use_motpy,
+                        n_motpy_tracks=self.n_motpy_tracks,
+                        use_tensorrt=self.use_tensorrt,
+                        total_frames=self.expected_video_length_frames,
+                    )
+                    # when completed, move to output file
+                    shutil.copy(temp_h5_path, output_h5_file)
+                    if os.path.exists(temp_h5_path):
+                        os.remove(temp_h5_path)
+                    # remove the log file
+                    if os.path.exists(log_file):
+                        os.remove(log_file)
+                except Exception as e:
+                    # Catch any exception and clean up
+                    print(f"Exception caught: {e}", file=sys.stderr)
+                    if os.path.exists(log_file):
+                        os.remove(log_file)
+                    if os.path.exists(temp_h5_path):
+                        os.remove(temp_h5_path)
+                    raise
+
         logger.info(f"Completed processing {len(self.all_videos)} videos")
-    
+
         # if the number of videos matches the number of h5 files, then we are done
         #  we do this in case there is a second job (e.g. local vs remote) working on the same data
         n_h5_predictions = len(list(self.output_directory_predictions.glob("*.h5")))
-        logger.info(
-            f"Completed {n_h5_predictions} out of {len(self.all_videos)} videos"
-        )
+        logger.info(f"Completed {n_h5_predictions} out of {len(self.all_videos)} videos")
         if n_h5_predictions == len(self.all_videos):
             # save a file completed.log in self.output_directory_predictions to indicate completion
             with open(self.output_directory_predictions / "completed.log", "w") as f:
@@ -222,7 +218,6 @@ class Inferencer2D:
             )
         else:
             logger.info("Not all videos completed, not writing completed.log")
-            
 
 
 def predict_video(
@@ -363,3 +358,22 @@ def predict_video(
         if use_motpy:
             h5f["detection_changes"] = detection_changes[: (frame_id + 1)]
             h5f["missing_detections"] = missing_detections[: (frame_id + 1)]
+
+
+def update_log_file_signal_handler(log_file):
+    """Utility to update the signal handler for the current log file.
+
+    This function exists in case the prediction is interrupted by e.g. tensorrt failing.
+    """
+
+    def handle_termination_signal():
+        print("Termination signal caught, deleting log file.", file=sys.stderr)
+        if os.path.exists(log_file):
+            os.remove(log_file)
+        logger.info("Termination signal caught, deleting log file.")
+        sys.exit(1)
+
+    # Register the signal handler for the current video log file
+    signal.signal(signal.SIGABRT, handle_termination_signal)
+    signal.signal(signal.SIGTERM, handle_termination_signal)
+    signal.signal(signal.SIGINT, handle_termination_signal)
